Separate `PyTransactionExecutor` to its own file.
