+++ crates/native_blockifier/src/py_transaction.rs
@@ -1,27 +1,15 @@
-use std::collections::{HashMap, HashSet};
 use std::convert::TryFrom;
 use std::sync::Arc;
 
 use blockifier::abi::constants::L1_HANDLER_VERSION;
-use blockifier::block_context::BlockContext;
-use blockifier::block_execution::pre_process_block;
 use blockifier::execution::contract_class::{ContractClassV0, ContractClassV1};
-use blockifier::state::cached_state::CachedState;
-use blockifier::state::state_api::{State, StateReader};
 use blockifier::transaction::account_transaction::AccountTransaction;
 use blockifier::transaction::objects::AccountTransactionContext;
 use blockifier::transaction::transaction_execution::Transaction;
 use blockifier::transaction::transaction_types::TransactionType;
-use blockifier::transaction::transactions::{
-    DeclareTransaction, ExecutableTransaction, L1HandlerTransaction,
-};
-use cairo_vm::vm::runners::cairo_runner::ExecutionResources as VmExecutionResources;
+use blockifier::transaction::transactions::{DeclareTransaction, L1HandlerTransaction};
 use num_bigint::BigUint;
-use ouroboros;
-use papyrus_storage::db::RO;
-use papyrus_storage::state::StateStorageReader;
 use pyo3::prelude::*;
-use starknet_api::block::{BlockHash, BlockNumber, BlockTimestamp};
 use starknet_api::core::{
     ClassHash, CompiledClassHash, ContractAddress, EntryPointSelector, Nonce,
 };
@@ -32,29 +20,8 @@ use starknet_api::transaction::{
     TransactionHash, TransactionSignature, TransactionVersion,
 };
 
-use crate::errors::{NativeBlockifierError, NativeBlockifierInputError, NativeBlockifierResult};
-use crate::papyrus_state::{PapyrusReader, PapyrusStateReader};
-use crate::py_state_diff::PyStateDiff;
-use crate::py_transaction_execution_info::{PyTransactionExecutionInfo, PyVmExecutionResources};
-use crate::py_utils::{biguint_to_felt, to_chain_id_enum, PyFelt};
-use crate::storage::Storage;
-
-fn py_attr<T>(obj: &PyAny, attr: &str) -> NativeBlockifierResult<T>
-where
-    T: for<'a> FromPyObject<'a>,
-    T: Clone,
-{
-    Ok(obj.getattr(attr)?.extract()?)
-}
-
-fn py_enum_name<T>(obj: &PyAny, attr: &str) -> NativeBlockifierResult<T>
-where
-    T: for<'a> FromPyObject<'a>,
-    T: Clone,
-    T: ToString,
-{
-    py_attr(obj.getattr(attr)?, "name")
-}
+use crate::errors::{NativeBlockifierInputError, NativeBlockifierResult};
+use crate::py_utils::{biguint_to_felt, py_attr};
 
 fn py_felt_attr(obj: &PyAny, attr: &str) -> NativeBlockifierResult<StarkFelt> {
     biguint_to_felt(py_attr(obj, attr)?)
@@ -80,41 +47,6 @@ pub fn py_account_data_context(tx: &PyAny) -> NativeBlockifierResult<AccountTran
         nonce,
         sender_address: ContractAddress::try_from(py_felt_attr(tx, "sender_address")?)?,
     })
-}
-
-pub fn py_block_context(
-    general_config: PyGeneralConfig,
-    block_info: &PyAny,
-) -> NativeBlockifierResult<BlockContext> {
-    let starknet_os_config = general_config.starknet_os_config;
-    let block_number = BlockNumber(py_attr(block_info, "block_number")?);
-    let block_context = BlockContext {
-        chain_id: to_chain_id_enum(starknet_os_config.chain_id)?,
-        block_number,
-        block_timestamp: BlockTimestamp(py_attr(block_info, "block_timestamp")?),
-        sequencer_address: ContractAddress::try_from(general_config.sequencer_address.0)?,
-        fee_token_address: ContractAddress::try_from(starknet_os_config.fee_token_address.0)?,
-        vm_resource_fee_cost: general_config.cairo_resource_fee_weights,
-        gas_price: py_attr(block_info, "gas_price")?,
-        invoke_tx_max_n_steps: general_config.invoke_tx_max_n_steps,
-        validate_max_n_steps: general_config.validate_max_n_steps,
-    };
-
-    Ok(block_context)
-}
-
-#[derive(FromPyObject)]
-pub struct PyGeneralConfig {
-    pub starknet_os_config: PyOsConfig,
-    pub sequencer_address: PyFelt,
-    pub cairo_resource_fee_weights: HashMap<String, f64>,
-    pub invoke_tx_max_n_steps: u32,
-    pub validate_max_n_steps: u32,
-}
-#[derive(FromPyObject)]
-pub struct PyOsConfig {
-    pub chain_id: BigUint,
-    pub fee_token_address: PyFelt,
 }
 
 pub fn py_declare(
@@ -271,255 +203,4 @@ pub fn py_tx(
         }
         _ => unimplemented!(),
     }
-}
-
-/// Wraps the transaction executor in an optional, to allow an explicit deallocation of it.
-/// The explicit deallocation is needed since PyO3 can't track lifetimes within Python.
-#[pyclass]
-pub struct PyTransactionExecutor {
-    pub executor: Option<PyTransactionExecutorInner>,
-}
-
-#[pymethods]
-impl PyTransactionExecutor {
-    #[new]
-    #[args(general_config, block_info, papyrus_storage)]
-    pub fn create(
-        general_config: PyGeneralConfig,
-        block_info: &PyAny,
-        papyrus_storage: &Storage,
-    ) -> NativeBlockifierResult<Self> {
-        log::debug!("Initializing Transaction Executor...");
-        let executor =
-            PyTransactionExecutorInner::create(general_config, block_info, papyrus_storage)?;
-        log::debug!("Initialized Transaction Executor.");
-
-        Ok(Self { executor: Some(executor) })
-    }
-
-    #[args(tx, raw_contract_class, enough_room_for_tx)]
-    pub fn execute(
-        &mut self,
-        tx: &PyAny,
-        raw_contract_class: Option<&str>,
-        // This is functools.partial(bouncer.add, tw_written=tx_written).
-        enough_room_for_tx: &PyAny,
-    ) -> NativeBlockifierResult<(Py<PyTransactionExecutionInfo>, PyVmExecutionResources)> {
-        self.executor().execute(tx, raw_contract_class, enough_room_for_tx)
-    }
-
-    pub fn finalize(&mut self) -> PyStateDiff {
-        log::debug!("Finalizing execution...");
-        let finalized_state = self.executor().finalize();
-        log::debug!("Finalized execution.");
-
-        finalized_state
-    }
-
-    pub fn close(&mut self) {
-        self.executor = None;
-    }
-
-    #[args(old_block_number_and_hash)]
-    pub fn pre_process_block(
-        &mut self,
-        old_block_number_and_hash: Option<(u64, PyFelt)>,
-    ) -> NativeBlockifierResult<()> {
-        self.executor().pre_process_block(old_block_number_and_hash)
-    }
-}
-
-impl PyTransactionExecutor {
-    fn executor(&mut self) -> &mut PyTransactionExecutorInner {
-        self.executor.as_mut().expect("Transaction executor should be initialized.")
-    }
-}
-
-// To access a field you must use `self.borrow_{field_name}()`.
-// Alternately, you can borrow the whole object using `self.with[_mut]()`.
-#[ouroboros::self_referencing]
-pub struct PyTransactionExecutorInner {
-    pub block_context: BlockContext,
-
-    // Maintained for counting purposes.
-    pub executed_class_hashes: HashSet<ClassHash>,
-
-    // State-related fields.
-    // Storage reader and transaction are kept merely for lifetime parameter referencing.
-    pub storage_reader: papyrus_storage::StorageReader,
-    #[borrows(storage_reader)]
-    #[covariant]
-    pub storage_tx: papyrus_storage::StorageTxn<'this, RO>,
-    #[borrows(storage_tx)]
-    #[covariant]
-    pub state: CachedState<PapyrusReader<'this>>,
-}
-
-impl PyTransactionExecutorInner {
-    pub fn create(
-        general_config: PyGeneralConfig,
-        block_info: &PyAny,
-        papyrus_storage: &Storage,
-    ) -> NativeBlockifierResult<Self> {
-        // Assumption: storage is aligned.
-        let reader = papyrus_storage.reader().clone();
-
-        let block_context = py_block_context(general_config, block_info)?;
-        build_tx_executor(block_context, reader)
-    }
-
-    /// Executes the given transaction on the state maintained by the executor.
-    /// Returns the execution trace, together with the compiled class hashes of executed classes
-    /// (used for counting purposes).
-    pub fn execute(
-        &mut self,
-        tx: &PyAny,
-        raw_contract_class: Option<&str>,
-        // This is functools.partial(bouncer.add, tw_written=tx_written).
-        enough_room_for_tx: &PyAny,
-    ) -> NativeBlockifierResult<(Py<PyTransactionExecutionInfo>, PyVmExecutionResources)> {
-        let tx_type: String = py_enum_name(tx, "tx_type")?;
-        let tx: Transaction = py_tx(&tx_type, tx, raw_contract_class)?;
-
-        let mut tx_executed_class_hashes = HashSet::<ClassHash>::new();
-        self.with_mut(|executor| {
-            let mut transactional_state = CachedState::create_transactional(executor.state);
-            let tx_execution_result = tx
-                .execute_raw(&mut transactional_state, executor.block_context)
-                .map_err(NativeBlockifierError::from);
-            let py_tx_execution_info = match tx_execution_result {
-                Ok(tx_execution_info) => {
-                    tx_executed_class_hashes = tx_execution_info.get_executed_class_hashes();
-                    Python::with_gil(|py| {
-                        // Allocate this instance on the Python heap.
-                        // This is necessary in order to pass a reference to it to the callback
-                        // (otherwise, if it were allocated on Rust's heap/stack, giving Python a
-                        // reference to the objects will not work).
-                        Py::new(py, PyTransactionExecutionInfo::from(tx_execution_info))
-                            .expect("Should be able to allocate on Python heap")
-                    })
-                }
-                Err(error) => {
-                    transactional_state.abort();
-                    return Err(error);
-                }
-            };
-
-            let has_enough_room_for_tx = Python::with_gil(|py| {
-                // Can be done because `py_tx_execution_info` is a `Py<PyTransactionExecutionInfo>`,
-                // hence is allocated on the Python heap.
-                let args = (py_tx_execution_info.borrow(py),);
-                enough_room_for_tx.call1(args) // Callback to Python code.
-            });
-
-            match has_enough_room_for_tx {
-                Ok(_) => {
-                    transactional_state.commit();
-                    let py_executed_compiled_class_hashes = get_casm_hash_calculation_resources(
-                        executor.state,
-                        executor.executed_class_hashes,
-                        &tx_executed_class_hashes,
-                    )?;
-                    Ok((py_tx_execution_info, py_executed_compiled_class_hashes))
-                }
-                // Unexpected error, abort and let caller know.
-                Err(error) if unexpected_callback_error(&error) => {
-                    transactional_state.abort();
-                    Err(error.into())
-                }
-                // Not enough room in batch, abort and let caller verify on its own.
-                Err(_not_enough_weight_error) => {
-                    transactional_state.abort();
-                    let py_executed_compiled_class_hashes = get_casm_hash_calculation_resources(
-                        executor.state,
-                        executor.executed_class_hashes,
-                        &tx_executed_class_hashes,
-                    )?;
-                    Ok((py_tx_execution_info, py_executed_compiled_class_hashes))
-                }
-            }
-        })
-    }
-
-    /// Returns the state diff resulting in executing transactions.
-    pub fn finalize(&mut self) -> PyStateDiff {
-        PyStateDiff::from(self.borrow_state().to_state_diff())
-    }
-
-    // Block pre-processing; see `block_execution::pre_process_block` documentation.
-    pub fn pre_process_block(
-        &mut self,
-        old_block_number_and_hash: Option<(u64, PyFelt)>,
-    ) -> NativeBlockifierResult<()> {
-        let old_block_number_and_hash = old_block_number_and_hash
-            .map(|(block_number, block_hash)| (BlockNumber(block_number), BlockHash(block_hash.0)));
-
-        self.with_mut(|executor| {
-            pre_process_block(executor.state, old_block_number_and_hash);
-        });
-
-        Ok(())
-    }
-}
-
-pub fn build_tx_executor(
-    block_context: BlockContext,
-    storage_reader: papyrus_storage::StorageReader,
-) -> NativeBlockifierResult<PyTransactionExecutorInner> {
-    // The following callbacks are required to capture the local lifetime parameter.
-    fn storage_tx_builder(
-        storage_reader: &papyrus_storage::StorageReader,
-    ) -> NativeBlockifierResult<papyrus_storage::StorageTxn<'_, RO>> {
-        Ok(storage_reader.begin_ro_txn()?)
-    }
-
-    fn state_builder<'a>(
-        storage_tx: &'a papyrus_storage::StorageTxn<'a, RO>,
-        block_number: BlockNumber,
-    ) -> NativeBlockifierResult<CachedState<PapyrusReader<'a>>> {
-        let state_reader = storage_tx.get_state_reader()?;
-        let state_reader = PapyrusStateReader::new(state_reader, block_number);
-        let papyrus_reader = PapyrusReader::new(storage_tx, state_reader);
-        Ok(CachedState::new(papyrus_reader))
-    }
-
-    let executed_class_hashes = HashSet::<ClassHash>::new();
-    let block_number = block_context.block_number;
-    // The builder struct below is implicitly created by `ouroboros`.
-    let py_tx_executor_builder = PyTransactionExecutorInnerTryBuilder {
-        block_context,
-        executed_class_hashes,
-        storage_reader,
-        storage_tx_builder,
-        state_builder: |storage_tx| state_builder(storage_tx, block_number),
-    };
-    py_tx_executor_builder.try_build()
-}
-
-fn unexpected_callback_error(error: &PyErr) -> bool {
-    let error_string = error.to_string();
-    !(error_string.contains("BatchFull") || error_string.contains("TransactionBiggerThanBatch"))
-}
-
-/// Returns the estimated VM resources for Casm hash calculation (done by the OS), of the newly
-/// executed classes by the current transaction.
-/// Also, updates the executed class hashes set, to be able to recognize the newly added ones.
-pub fn get_casm_hash_calculation_resources(
-    state: &mut CachedState<PapyrusReader<'_>>,
-    executed_class_hashes: &mut HashSet<ClassHash>,
-    tx_executed_class_hashes: &HashSet<ClassHash>,
-) -> NativeBlockifierResult<PyVmExecutionResources> {
-    let newly_executed_class_hashes: HashSet<&ClassHash> = tx_executed_class_hashes
-        .iter()
-        .filter(|&class_hash| executed_class_hashes.insert(*class_hash))
-        .collect();
-
-    let mut casm_hash_computation_resources = VmExecutionResources::default();
-
-    for class_hash in newly_executed_class_hashes {
-        let class = state.get_compiled_contract_class(class_hash)?;
-        casm_hash_computation_resources += &class.estimate_casm_hash_computation_resources();
-    }
-
-    Ok(PyVmExecutionResources::from(casm_hash_computation_resources))
 }
